+++ +helics/HELICS_TRUE.m
@@ -4,7 +4,7 @@ indicator used for a true response
 function v = HELICS_TRUE()
 	persistent vInitialized;
 	if isempty(vInitialized)
-		vInitialized = 1;
+		vInitialized = int32(1);
 	end
 	v = vInitialized;
 end
